some refactoring before merge
