+++ src/data/en/index.ts
@@ -761,7 +761,6 @@ export const loggedInData = {
             'To make the content draggable, convert them for the new editor:',
           oneColumnLayout: 'One-column layout',
           multimediaTitle: 'Multimedia content associated with text',
-          chooseLayout: 'Choose layout',
         },
         multimedia: {
           title: 'Multimedia content associated with text',
